requirements.txt
requests==2.20.1
kafka-python==2.0.2